feat: Update npm dependencies and TypeScript version
