+++ src/app/(aplicacao)/@modal/participar/page.tsx
@@ -1,0 +1,5 @@
+import { ModalMenu } from '../ModalMenu';
+
+export default function Page() {
+  return <ModalMenu variant="key" />;
+}
