+++ src/app/(aplicacao)/@modal/validar/page.tsx
@@ -1,0 +1,5 @@
+import { ModalMenu } from '../ModalMenu';
+
+export default function Page() {
+  return <ModalMenu variant="hash" />;
+}
