+++ src/app/(aplicacao)/historico/page.tsx
@@ -11,70 +11,70 @@ const cards = [
     description: 'Descrição da enquete 1',
     variant: 'VOTAÇÃO' as Variant,
     hashtags: ['tag1', 'tag2'],
-    imageUrl: 'url1',
+    imageUrl: 'https://picsum.photos/300/200',
   },
   {
     title: 'Enquete 2',
     description: 'Descrição da enquete 2',
     variant: 'ENQUETE' as Variant,
     hashtags: ['tag1', 'tag2'],
-    imageUrl: 'url2',
+    imageUrl: 'https://picsum.photos/300/200',
   },
   {
     title: 'Enquete 3',
     description: 'Descrição da enquete 3',
     variant: 'VOTAÇÃO' as Variant,
     hashtags: ['tag1', 'tag2'],
-    imageUrl: 'url3',
+    imageUrl: 'https://picsum.photos/300/200',
   },
   {
     title: 'Enquete 4',
     description: 'Descrição da enquete 4',
     variant: 'ENQUETE' as Variant,
     hashtags: ['tag1', 'tag2'],
-    imageUrl: 'url4',
+    imageUrl: 'https://picsum.photos/300/200',
   },
   {
     title: 'Enquete 5',
     description: 'Descrição da enquete 5',
     variant: 'VOTAÇÃO' as Variant,
     hashtags: ['tag1', 'tag2'],
-    imageUrl: '5',
+    imageUrl: 'https://picsum.photos/300/200',
   },
   {
     title: 'Enquete 6',
     description: 'Descrição da enquete 6',
     variant: 'ENQUETE' as Variant,
     hashtags: ['tag1', 'tag2'],
-    imageUrl: 'url6',
+    imageUrl: 'https://picsum.photos/300/200',
   },
   {
     title: 'Enquete 7',
     description: 'Descrição da enquete 7',
     variant: 'VOTAÇÃO' as Variant,
     hashtags: ['tag1', 'tag2'],
-    imageUrl: 'url7',
+    imageUrl: 'https://picsum.photos/300/200',
   },
   {
     title: 'Enquete 8',
     description: 'Descrição da enquete 8',
     variant: 'ENQUETE' as Variant,
     hashtags: ['tag1', 'tag2'],
-    imageUrl: 'url8',
+    imageUrl: 'https://picsum.photos/300/200',
   },
   {
     title: 'Enquete 9',
     description: 'Descrição da enquete 9',
     variant: 'VOTAÇÃO' as Variant,
     hashtags: ['tag1', 'tag2'],
-    imageUrl: 'url9',
+    imageUrl: 'https://picsum.photos/300/200',
   },
   {
     title: 'Enquete 10',
     description: 'Descrição da enquete 10',
     variant: 'ENQUETE' as Variant,
     hashtags: ['tag1', 'tag2'],
-    imageUrl: 'url10',
+    imageUrl: 'https://picsum.photos/300/200',
   },
 ];
