+++ src/app/(auth)/login/page.tsx
@@ -4,7 +4,7 @@ import InputCustom from '@/components/InputCustom/InputCustom';
 import Butao from '@/components/buttons/button';
 import { Metadata } from 'next';
 import Link from 'next/link';
-import { axiosClient, setCookie } from '../../../../lib';
+import { entrar, setCookie } from '@/lib/auth';
 import clsx from 'clsx';
 import { useRouter } from 'next/navigation';
 
@@ -21,23 +21,18 @@ export default function Login() {
   const [showAlerta, setNewShowAlerta] = useState(false);
   const router = useRouter();
 
-  const sendSignIn = (email: any, senha: any) => {
-    axiosClient
-      .post('/auth/entrar', {
-        email: email,
-        senha: senha,
-      })
-      .then((res) => {
-        console.log('opa');
-        setCookie('accessToken', res.data.accessToken);
-        router.push('/inicio');
-      })
-      .catch((err) => {
-        setNewShowAlerta(true);
-        setTimeout(() => {
-          setNewShowAlerta(false);
-        }, 3000);
-      });
+  const sendSignIn = async (email: string, senha: string) => {
+    const token = await entrar(email, senha);
+
+    if (token) {
+      setCookie('accessToken', token);
+      router.push('/inicio');
+    } else {
+      setNewShowAlerta(true);
+      setTimeout(() => {
+        setNewShowAlerta(false);
+      }, 3000);
+    }
   };
 
   return (
